+++ packages/lib/modules/eclp/helpers/gyroECLPMath.ts
