+++ packages/lib/modules/eclp/helpers/gyroEMath.ts
