+++ packages/lib/modules/eclp/helpers/gyroEMathFunctions.ts
