+++ packages/lib/modules/eclp/helpers/gyroEMathHelpers.ts
