+++ packages/lib/modules/eclp/helpers/gyroSignedFixedPoint.ts
