+++ packages/lib/modules/pool/PoolDetail/PoolStats/PoolCharts/useEclpPoolChart.tsx
@@ -1,4 +1,4 @@
-import { useGetECLPLiquidityProfile } from '@repo/lib/modules/eclp/useGetECLPLiquidityProfile'
+import { useGetECLPLiquidityProfile } from '@repo/lib/modules/eclp/hooks/useGetECLPLiquidityProfile'
 import { bn, fNum } from '@repo/lib/shared/utils/numbers'
 import { usePool } from '../../../PoolProvider'
 import { ColorMode, useTheme as useChakraTheme } from '@chakra-ui/react'
